requirements.txt
numpy
pandas
streamlit==1.24.0
pillow
gdown
plotly
tensorflow
tensorrt